backend/requirements.txt
fastapi
uvicorn[standard]
websockets
mediapipe
ultralytics
Pillow
pillow-heif
ruff
pytest
